Parse star, plus and optional
Minor tweaks to make this happen.
RegexParser::get_next_token needed to report an error
while continuing execution. It can be okay to ask for
the next token, even though we're at the end of the
stream. For example when trying one production, before
backtracking.
As it stands now, it doesn't show the nice error message
ever, so I need to get that back in the right circumstances.