AST/visitors: Parser builds an AST while parsing
The AST also has a visitor interface, so we can implement different
visitors to process it.
This commit implements a printing visitor.

After this commit, we should be able to improve our testing dramatically
by actually inspecting the ASTs and verifying that they look like we
imagine, instead of relying solely on the status flags of the parser