parser: add variadic template argument for error msg
Not sure how much i will use it, or if it is an ergonomic implementation,
but it seemed like a good idea at the time.
Allows me to state which types of tokens i expect and print an
appropriate error message. Previously, i could only state one expected
token, which wouldnt neccessarily be the whole truth.